+++ tests/testthat/test-helpers.R
@@ -175,7 +175,7 @@ attributes(df1)$label <- "Test data 1"
 attributes(df1$mpg)$label <- "Col label 1"
 attributes(df1$type)$label <- "Col label 2"
 
-test_that("generate_choices() generates meaningful choices for datasets and columns to be used in the corresponding dropdown menues" %>% # nolint
+test_that("generate_choices() generates meaningful choices for datasets and columns to be used in the corresponding dropdown menus" |> # nolint
   vdoc[["add_spec"]](
     c(
       specs$listings_label,
@@ -318,10 +318,9 @@ test_that("set_up_datatable() returns correct column names, row names, and pagin
   attributes(df$A)$label <- "Label A"
   attributes(df$C)$label <- "Label C"
 
-  selected_cols <- c("A", "B", "C")
   pagination <- NULL
 
-  actual <- set_up_datatable(df, selected_cols, pagination)
+  actual <- set_up_datatable(df, pagination)
 
   expected <- list(
     col_names = c("A [Label A]", "B [No label]", "C [Label C]"),
@@ -338,10 +337,9 @@ test_that("set_up_datatable() automatically activates pagination for large datas
     C = sample(c("a", "b", "c"), 1001, replace = TRUE)
   )
 
-  selected_cols <- c("A", "B", "C")
   pagination <- NULL
 
-  actual <- set_up_datatable(df, selected_cols, pagination)
+  actual <- set_up_datatable(df, pagination)
 
   expect_true(actual$paging)
 })
@@ -353,10 +351,9 @@ test_that("set_up_datatable() automatically deactivates pagination for small dat
     C = sample(c("a", "b", "c"), 100, replace = TRUE)
   )
 
-  selected_cols <- c("A", "B", "C")
   pagination <- NULL
 
-  actual <- set_up_datatable(df, selected_cols, pagination)
+  actual <- set_up_datatable(df, pagination)
 
   expect_false(actual$paging)
 })