+++ NEWS.md
@@ -1,6 +1,11 @@
-# dv.listings 4.1.1-9001
+# dv.listings 4.1.2-9000
 
-- Adapt checkmate calls to allow default values for datasets that are not present
+- Includes the experimental review functionality
+
+# dv.listings 4.1.1-9003
+
+- Adapt fill_default_vars to keep specified default values for datasets which are not present
+- Adapt checkmate calls to allow default values for datasets which are not present
 - Top buttons are reordered in a single line to avoid consuming excessive space
 
 # dv.listings 4.1.0